Adds a translation loader for plugins
Plugins are different, they have a type or owner. The translations are
installed to a owner subfolder. The logic that applies to libs and apps
fails with plugins.
We search for the translations in /pluginnOwner/pluginName relative to the
base search path provided by getSearchPaths();
Ex: /usr/share/lxqt-qt5/translations/lxqt-panel/quicklaunch/
lxqt-panel is the type and quicklaunch is the plugin.