Adds Qt5 automatic load of self translations
loadSelfTranslation() will be called from the QCoreApplication constructor.
This way we can load the translation without requiring the application to
call it. It piles up, we can have all libraries translations self loaded
with Q_CORE_STARTUP_FUNCTION macro. It works for Qt 5.1 and later.

Sadly such a macro isn't available for Qt4.