Don't update (extract) translations by default.
If they are enabled by default in each build there will be, probably,
new .ts files. And they will not be commited. They are meant to be commited
in an Update Translations commit. They will be left behind and in the next
git fetch/git submodule update, they will have to be reset.

The proper solution is to update at an "regular" interval and have an
Translator call (freeze). More development savvy translators can also
update the translations themselves.