Adds the LXQtranslateTs cmake module
We now have the ability to choose when to update (extract) the
translations from the sources files.

To update the translations set the CMake command line option
UPDATE_TRANSLATIONS to true. Ex: cmake -DUPDATE_TRASNLATIONS=Yes ..
The translations will be extracted and also compiled.
To return to the compilation only mode just set the UPDATE_TRANSLATIONS
option to false. Ex: cmake -DCMAKE_UPDATE_TRANSLATIONS=OFF ..

Warning: When UPDATE_TRANSLATIONS is enabled a clean command will also
clean the generated .ts files. The CLEAN_NO_CUSTOM property should solve
that problem. But it didn't, at least for now. So, after updating the
translations, disable UPDATE_TRANSLATIONS right away.

The LxQtTranslateTs module is part of the lxqt find_package(). The module
path is appended to CMAKE_MODULE_PATH, so the user only needs to include
the modules.