Adds StartOptions to the SingleApplication class
By default (ExitOnDBusFailure) if an instance can't connect to the D-Bus
session bus, that instance calls ::exit(1). Not even the first instance
will run. Connecting to the D-Bus session bus is an condition to
guarantee that only one instance will run.

If an user wants to allow an application to run without D-Bus, it must
use the NoExitOnDBusFailure option.

ExitOnDBusFailure is the default.