lxqt-config.cmake.in: Use find_dependency()
Replaces find_package() with find_dependency(). From the documentation:

"find_dependency() wraps a find_package() call for a package dependency.
It is designed to be used in a <package>Config.cmake file, and it forwards
the correct parameters for EXACT, QUIET and REQUIRED which were passed to
the original find_package() call. It also sets an informative diagnostic
message if the dependency could not be found."
CMake v3.0.2 or higher or required.