LXQtTheme: Remove (outdated) "singleclick" qss assembly
Assembly of the "activate-on-singleclick" style hint into style sheet is outdated. The *desktop.conf* confiugration
is not used any more (it is a remnant from the razor days). *lxqt-config-input* stores this configuration in
"lxqt.conf" settings file (key "single_click_activate") and this value is used in our *lxqt-qtplugin* (QPA platform theme)