lxqtapplication: Avoid wrong signal handling
Add the application object as receiver to QObject::connect() -> to
remove the connection automaticaly when the application object is
destructed.
W/o the automatic connection removal we could end in SEGFAULT if UNIX
signal is delivered after the QApplication destruction.

(this was a corner case with almost zero probablity, but now the UNIX
signal handling should be 100% correct)