Position, alignment and size fixes for HTML delegate
Closes https://github.com/lxqt/lxqt/issues/1550

Texts and icons are vertically centered and item heights are correctly calculated in terms of wrapped text heights and icon heights. Unconditional using of `QTextDocument::adjustSize()` was wrong (actually, it could be removed). Moreover, `QStyle::drawControl()` should not be used for drawing the panel.

Please also note that Qt has an HTML bug, which is explained in a code comment.