Fix compilation error on the Hurd.
This commit disables the Linux backend on the Hurd because it fails to
compile on Debian GNU/Hurd and does not seem to be useful on the Hurd
(the code contains paths which do not exist on Hurd).

The build failure was caused by an unconditional use of PATH_MAX, which
is not defined on the Hurd.