Backlight: Fix LinuxBackend backlightChanged signal
VirtualBackEnd defined the backlightChanged signal. But the LinuxBackend
inherits VirtualBackEnd and defines the signal again, overriding it.

LXQt::Backlight correctly uses the VirtualBackEnd::backlightChanged signal:
connect(m_backend, &VirtualBackEnd::backlightChanged,
    this, &Backlight::backlightChangedSlot);

and that implies that for the LinuxBackend case the backlightChangedSlot
slot is never called. The signal source is LinuxBackend not VirtualBackEnd.

The fix is simple, just delete the LinuxBackend backlightChanged signal.

A user using the LXQt::Backlight would never get the backlightChanged
signal.